Move PropertyInjector into astmodel
