+++ hack/generator/pkg/astmodel/property_injector.go
@@ -3,21 +3,19 @@
  * Licensed under the MIT license.
  */
 
-package storage
-
-import "github.com/Azure/azure-service-operator/hack/generator/pkg/astmodel"
+package astmodel
 
 // PropertyInjector is a utility for injecting property definitions into resources and objects
 type PropertyInjector struct {
 	// visitor is used to do the actual injection
-	visitor astmodel.TypeVisitor
+	visitor TypeVisitor
 }
 
 // NewPropertyInjector creates a new property injector for modifying resources and objects
 func NewPropertyInjector() *PropertyInjector {
 	result := &PropertyInjector{}
 
-	result.visitor = astmodel.TypeVisitorBuilder{
+	result.visitor = TypeVisitorBuilder{
 		VisitObjectType:   result.injectPropertyIntoObject,
 		VisitResourceType: result.injectPropertyIntoResource,
 	}.Build()
@@ -26,20 +24,20 @@ func NewPropertyInjector() *PropertyInjector {
 }
 
 // Inject modifies the passed type definition by injecting the passed property
-func (pi *PropertyInjector) Inject(def astmodel.TypeDefinition, prop *astmodel.PropertyDefinition) (astmodel.TypeDefinition, error) {
+func (pi *PropertyInjector) Inject(def TypeDefinition, prop *PropertyDefinition) (TypeDefinition, error) {
 	return pi.visitor.VisitDefinition(def, prop)
 }
 
 // injectPropertyIntoObject takes the property provided as a context and includes it on the provided object type
 func (pi *PropertyInjector) injectPropertyIntoObject(
-	_ *astmodel.TypeVisitor, ot *astmodel.ObjectType, ctx interface{}) (astmodel.Type, error) {
-	prop := ctx.(*astmodel.PropertyDefinition)
+	_ *TypeVisitor, ot *ObjectType, ctx interface{}) (Type, error) {
+	prop := ctx.(*PropertyDefinition)
 	return ot.WithProperty(prop), nil
 }
 
 // injectPropertyIntoResource takes the property  provided as a context and includes it on the provided resource type
 func (pi *PropertyInjector) injectPropertyIntoResource(
-	_ *astmodel.TypeVisitor, rt *astmodel.ResourceType, ctx interface{}) (astmodel.Type, error) {
-	prop := ctx.(*astmodel.PropertyDefinition)
+	_ *TypeVisitor, rt *ResourceType, ctx interface{}) (Type, error) {
+	prop := ctx.(*PropertyDefinition)
 	return rt.WithProperty(prop), nil
 }
